update fixture structure and mark iron converter tests
