+++ h2integrate/converters/iron/rosner/cost_coeffs.csv
@@ -1,27 +1,27 @@
 Name,Type,Coeff,Unit,h2_dri_eaf,h2_dri,ng_dri_eaf,ng_dri,h2_eaf,ng_eaf
 Dollar Year,all,-,-,2022.0,2022.0,2022.0,2022.0,2022.0,2022.0
-EAF & Casting,capital,lin,$,352191.5240169328,9.999999999999783e-11,348441.82271277835,9.999999999999996e-11,352191.5240169328,348441.82271277835
-Shaft Furnace,capital,lin,$,489.68060552071756,498.4011587515349,12706.35540748921,12706.35540748921,9.999999999999783e-11,9.999999999999996e-11
-Reformer,capital,lin,$,0.0,0.0,12585.824569411547,12585.824569411547,0.0,9.999999999999996e-11
-Recycle Compressor,capital,lin,$,0.0,0.0,956.7744259199391,956.7744259199391,0.0,9.999999999999996e-11
-Oxygen Supply,capital,lin,$,1715.21508561117,1364.556123425206,1204.989085989581,1024.5621894371002,393.8307342018522,201.3153857876692
-H2 Pre-heating,capital,lin,$,45.69122789208198,45.69122789208198,0.0,0.0,9.999999999999783e-11,0.0
-Cooling Tower,capital,lin,$,2513.0831352600603,2349.15226276371,1790.4037306002654,1774.6459078322403,195.45026173489327,428.90723393800783
-Piping,capital,lin,$,11815.72718570437,172.83401368816206,25651.756587058582,3917.34823234076,49970.89662834664,51294.534191987485
-Electrical & Instrumentation,capital,lin,$,7877.151460413984,115.2226758287703,17101.171070383258,2611.5654909389123,33313.93113452472,34196.35613320244
-"Buildings, Storage, Water Service",capital,lin,$,1097.818759618821,630.531300861788,1077.3955282629602,618.8012346570356,467.2874587570587,458.59429360590775
-Other Miscellaneous Cost,capital,lin,$,7877.151460413984,115.2226758287703,17101.171070383258,2611.5654909389123,32124.80020341839,34196.35613320244
-EAF & Casting,capital,exp,-,0.4559999999420573,1.3155100341401327e-15,0.45599999989917783,-3.9460856373682605e-16,0.4559999999420573,0.45599999989917783
-Shaft Furnace,capital,exp,-,0.8874110806752277,0.8862693772994416,0.654083508755392,0.654083508755392,1.3155100341401327e-15,-3.9460856373682605e-16
-Reformer,capital,exp,-,0.0,0.0,0.6504523630280986,0.6504523630280986,0.0,-3.9460856373682605e-16
-Recycle Compressor,capital,exp,-,0.0,0.0,0.7100000000012126,0.7100000000012126,0.0,-3.9460856373682605e-16
-Oxygen Supply,capital,exp,-,0.6457441946722564,0.634272491606804,0.6448555996459648,0.6370664161269146,0.6699999996017935,0.6699999991468175
-H2 Pre-heating,capital,exp,-,0.8656365854575331,0.8656365854575331,0.0,0.0,1.3155100341401327e-15,0.0
-Cooling Tower,capital,exp,-,0.6332532740097354,0.6286978709250873,0.6302820063981899,0.6308163319151914,0.6659797264431847,0.25999986981600937
-Piping,capital,exp,-,0.5998309612457874,0.8331608480295299,0.5641056316548689,0.6551154484929355,0.461960765339552,0.45807183935606205
-Electrical & Instrumentation,capital,exp,-,0.599830961215121,0.8331608479997342,0.5641056316058728,0.6551154484197809,0.4619607652382664,0.4580718393497376
-"Buildings, Storage, Water Service",capital,exp,-,0.7999999998942431,0.7999999999824277,0.7999999998654156,0.7999999999666229,0.7999999997752477,0.7999999997288553
-Other Miscellaneous Cost,capital,exp,-,0.599830961215121,0.8331608479997342,0.5641056316058728,0.6551154484197809,0.46389753648982157,0.4580718393497376
+EAF & Casting,capital,lin,$,352191.5240169347,1.0000000000000102e-10,348441.82271277584,9.999999999999676e-11,352191.5240169347,348441.82271277584
+Shaft Furnace,capital,lin,$,489.68060552073234,498.40115875154686,12706.355407489007,12706.355407489007,1.0000000000000102e-10,9.999999999999676e-11
+Reformer,capital,lin,$,0.0,0.0,12585.82456941159,12585.82456941159,0.0,9.999999999999676e-11
+Recycle Compressor,capital,lin,$,0.0,0.0,956.7744259199475,956.7744259199475,0.0,9.999999999999676e-11
+Oxygen Supply,capital,lin,$,1715.2150856112128,1364.5561234251575,1204.9890859895447,1024.5621894370765,393.830734201856,201.31538578766632
+H2 Pre-heating,capital,lin,$,45.691227892082026,45.691227892082026,0.0,0.0,1.0000000000000102e-10,0.0
+Cooling Tower,capital,lin,$,2513.0831352600294,2349.1522627636514,1790.4037306002654,1774.645907832289,195.4502617348917,428.90723393801625
+Piping,capital,lin,$,11815.727185704853,172.83401368816575,25651.75658705849,3917.3482323407948,49970.89662834815,51294.534191985025
+Electrical & Instrumentation,capital,lin,$,7877.151460413984,115.2226758287706,17101.171070383836,2611.565490938924,33313.93113452436,34196.35613320044
+"Buildings, Storage, Water Service",capital,lin,$,1097.8187596188552,630.5313008617601,1077.3955282629754,618.8012346570307,467.28745875705494,458.59429360591264
+Other Miscellaneous Cost,capital,lin,$,7877.151460413984,115.2226758287706,17101.171070383836,2611.565490938924,32124.80020341787,34196.35613320044
+EAF & Casting,capital,exp,-,0.45599999994205653,-1.0107865114766218e-15,0.4559999998991786,2.45988384474349e-15,0.45599999994205653,0.4559999998991786
+Shaft Furnace,capital,exp,-,0.8874110806752252,0.8862693772994397,0.654083508755393,0.654083508755393,-1.0107865114766218e-15,2.45988384474349e-15
+Reformer,capital,exp,-,0.0,0.0,0.650452363028098,0.650452363028098,0.0,2.45988384474349e-15
+Recycle Compressor,capital,exp,-,0.0,0.0,0.7100000000012116,0.7100000000012116,0.0,2.45988384474349e-15
+Oxygen Supply,capital,exp,-,0.6457441946722542,0.6342724916068068,0.6448555996459671,0.6370664161269166,0.6699999996017927,0.6699999991468184
+H2 Pre-heating,capital,exp,-,0.865636585457533,0.865636585457533,0.0,0.0,-1.0107865114766218e-15,0.0
+Cooling Tower,capital,exp,-,0.6332532740097363,0.6286978709250893,0.6302820063981899,0.6308163319151892,0.6659797264431855,0.25999986981600776
+Piping,capital,exp,-,0.5998309612457838,0.8331608480295282,0.5641056316548687,0.6551154484929348,0.4619607653395496,0.45807183935606577
+Electrical & Instrumentation,capital,exp,-,0.5998309612151209,0.8331608479997339,0.5641056316058699,0.6551154484197805,0.4619607652382672,0.4580718393497419
+"Buildings, Storage, Water Service",capital,exp,-,0.7999999998942408,0.799999999982431,0.7999999998654146,0.7999999999666235,0.7999999997752482,0.7999999997288545
+Other Miscellaneous Cost,capital,exp,-,0.5998309612151209,0.8331608479997339,0.5641056316058699,0.6551154484197805,0.4638975364898231,0.4580718393497419
 Preproduction,owner,lin,frac of TPC,0.02,0.02,0.02,0.02,0.02,0.02
 Spare Parts,owner,lin,frac of TPC,0.005,0.005,0.005,0.005,0.005,0.005
 "Initial Catalyst, Sorbent & Chemicals",owner,lin,frac of TPC,0.0,0.0,0.250835587,0.250835587,0.0,0.250835587
