+++ h2integrate/converters/iron/test/test_iron_plant.py
@@ -2,8 +2,6 @@ import pytest
 import openmdao.api as om
 from pytest import fixture
 
-from h2integrate import EXAMPLE_DIR
-from h2integrate.core.inputs.validation import load_driver_yaml
 from h2integrate.converters.iron.iron_plant import (
     IronPlantCostComponent,
     IronPlantPerformanceComponent,
@@ -40,32 +38,7 @@ def iron_dri_config_rosner_ng():
     return tech_config
 
 
-@fixture
-def plant_config():
-    plant_config = {
-        "plant": {
-            "plant_life": 30,
-            "simulation": {
-                "n_timesteps": 8760,
-                "dt": 3600,
-            },
-        },
-        "finance_parameters": {
-            "cost_adjustment_parameters": {
-                "cost_year_adjustment_inflation": 0.025,
-                "target_dollar_year": 2022,
-            }
-        },
-    }
-    return plant_config
-
-
-@fixture
-def driver_config():
-    driver_config = load_driver_yaml(EXAMPLE_DIR / "21_iron_mn_to_il" / "driver_config.yaml")
-    return driver_config
-
-
+@pytest.mark.regression
 def test_baseline_iron_dri_costs_rosner_ng(
     plant_config, driver_config, iron_dri_config_rosner_ng, subtests
 ):
@@ -103,6 +76,7 @@ def test_baseline_iron_dri_costs_rosner_ng(
         assert pytest.approx(prob.get_val("dri_cost.VarOpEx")[0], rel=1e-6) == expected_var_om
 
 
+@pytest.mark.regression
 def test_baseline_iron_dri_costs_rosner_h2(
     plant_config, driver_config, iron_dri_config_rosner_ng, subtests
 ):
