+++ h2integrate/converters/iron/test/test_martin_mine.py
@@ -3,8 +3,6 @@ import pytest
 import openmdao.api as om
 from pytest import fixture
 
-from h2integrate import EXAMPLE_DIR
-from h2integrate.core.inputs.validation import load_driver_yaml
 from h2integrate.converters.iron.martin_mine_cost_model import MartinIronMineCostComponent
 from h2integrate.converters.iron.martin_mine_perf_model import MartinIronMinePerformanceComponent
 
@@ -24,32 +22,7 @@ def iron_ore_config_martin_om():
     return tech_config
 
 
-@fixture
-def plant_config():
-    plant_config = {
-        "plant": {
-            "plant_life": 30,
-            "simulation": {
-                "n_timesteps": 8760,
-                "dt": 3600,
-            },
-        },
-        "finance_parameters": {
-            "cost_adjustment_parameters": {
-                "cost_year_adjustment_inflation": 0.025,
-                "target_dollar_year": 2022,
-            }
-        },
-    }
-    return plant_config
-
-
-@fixture
-def driver_config():
-    driver_config = load_driver_yaml(EXAMPLE_DIR / "21_iron_mn_to_il" / "driver_config.yaml")
-    return driver_config
-
-
+@pytest.mark.regression
 def test_iron_mine_performance_outputs(
     plant_config, driver_config, iron_ore_config_martin_om, subtests
 ):
@@ -149,6 +122,7 @@ def test_iron_mine_performance_outputs(
         assert np.all(prob.get_val("comp.replacement_schedule", units="unitless") == 0)
 
 
+@pytest.mark.regression
 def test_baseline_iron_ore_costs(plant_config, driver_config, iron_ore_config_martin_om, subtests):
     martin_ore_capex = 1221599018.626594
     martin_ore_fixed_om = 0.0
